Uploading a new version will properly store both versions.
There was a problem where the repository didn't want to update
the registry until after the storage had successfully uploaded
the package. However, the storage needed the updated information
in order to figure out the package filename.

The fix is to clone the registry entry, updated it and then swap in
the new one after the storage has successfully uploaded the package.